Close h2 database after each access again
The JdbcH2DatastoreProvider is currently designed to lazily open the h2 database on
first access and close it when the JVM shuts down.
But when using the Gradle daemon, the JVM won't shut down immediately, so
the database will be kept open.
Solution: Open and close the database on each access.